add segues to pass data to detailsVC
