--- Travl.iOS/View/DiscoverCell.swift
@@ -1,24 +1,0 @@
-//
-//  DiscoverCell.swift
-//  Travl.iOS
-//
-//  Created by Ikmal Azman on 11/08/2021.
-//
-
-import UIKit
-
-class DiscoverCell: UICollectionViewCell {
-    
-    @IBOutlet weak var locationImage: UIImageView!
-    @IBOutlet weak var locationLabel: UILabel!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        configureCell()
-    }
-    
-    func configureCell() {
-        locationImage.layer.cornerRadius = 15
-    }
-    
-}
